v 0.79.0
Detect Hack :
- Use default checksum check.
- Handle when no results.

+++ bin/detecthack.sh
@@ -15,37 +15,44 @@ _WPUDHK_DIR="tmpwp${_WPUDHK_RAND}";
 ## Questions
 ###################################
 
-_WPUDHK_COMPARE_WP=$(bashutilities_get_yn "- Compare core code to a fresh WordPress install?" 'y');
-_WPUDHK_COMPARE_PLUG=$(bashutilities_get_yn "- Compare each plugin code to a fresh plugin install?" "${_WPUDHK_COMPARE_WP}");
+_WPUDHK_COMPARE_WP=$(bashutilities_get_yn "- Compare core code to a fresh WordPress install instead of checksums?" 'y');
+_WPUDHK_COMPARE_PLUG='n';
+if [[ "${_WPUDHK_COMPARE_WP}" == 'y' ]];then
+    _WPUDHK_COMPARE_PLUG=$(bashutilities_get_yn "- Compare each plugin code to a fresh plugin install?" "${_WPUDHK_COMPARE_WP}");
+else
+    _WPUDHK_DIR=".";
+fi;
 
 ###################################
 ## Loading WordPress
 ###################################
 
-mkdir "${_WPUDHK_DIR}";
-
-# Copy wp-config to allow plugin install to work
-cp wp-config.php "${_WPUDHK_DIR}";
-
-# Extract version
-_CURRENT_WORDPRESS=$(_WPCLICOMMAND core version);
-
-# Extract local package
-_WPUDHK_LOCAL_PACKAGE=$(bashutilities_search_extract_file "\$wp_local_package = '" "';" "wp-includes/version.php");
-if [[ "${_WPUDHK_LOCAL_PACKAGE}" == '' ]];then
-    _WPUDHK_LOCAL_PACKAGE='en_US';
-fi;
-
-# Downloading test version
 if [[ "${_WPUDHK_COMPARE_WP}" == 'y' ]];then
-echo "# Downloading core WordPress to have a clean base to compare";
-_WPCLICOMMAND core download \
-    --quiet \
-    --skip-plugins \
-    --skip-themes \
-    --locale="${_WPUDHK_LOCAL_PACKAGE}" \
-    --path="${_WPUDHK_DIR}" \
-    --version="${_CURRENT_WORDPRESS}";
+    mkdir "${_WPUDHK_DIR}";
+
+    # Copy wp-config to allow plugin install to work
+    cp wp-config.php "${_WPUDHK_DIR}";
+
+    # Extract version
+    _CURRENT_WORDPRESS=$(_WPCLICOMMAND core version);
+
+    # Extract local package
+    _WPUDHK_LOCAL_PACKAGE=$(bashutilities_search_extract_file "\$wp_local_package = '" "';" "wp-includes/version.php");
+    if [[ "${_WPUDHK_LOCAL_PACKAGE}" == '' ]];then
+        _WPUDHK_LOCAL_PACKAGE='en_US';
+    fi;
+
+    # Downloading test version
+    echo "# Downloading core WordPress to have a clean base to compare";
+    _WPCLICOMMAND core download \
+        --quiet \
+        --skip-plugins \
+        --skip-themes \
+        --locale="${_WPUDHK_LOCAL_PACKAGE}" \
+        --path="${_WPUDHK_DIR}" \
+        --version="${_CURRENT_WORDPRESS}";
+else
+    _WPCLICOMMAND core verify-checksums;
 fi;
 
 if [[ "${_WPUDHK_COMPARE_PLUG}" == 'y' ]];then
@@ -65,6 +72,8 @@ for line in ${_WPPLUGINSTMPLIST}; do
             --quiet;
     fi;
 done
+else
+    _WPCLICOMMAND plugin verify-checksums --all;
 fi;
 
 ###################################
@@ -91,5 +100,6 @@ $_PHP_COMMAND "${_WPUDHK_FILE}" --dir="${_WPUDHK_DIR}";
 rm "${_WPUDHK_FILE}";
 
 # TMP WordPress
-rm -r "${_WPUDHK_DIR}";
-
+if [[ "${_WPUDHK_COMPARE_WP}" == 'y' ]];then
+    rm -r "${_WPUDHK_DIR}";
+fi;
